Move LEDs to its own file.
